+++ src/test/kotlin/graphql/operations/Fragments.kt
@@ -135,6 +135,37 @@ const val AUDIO_MESSAGE_FRAGMENT = """
     }
 """
 
+const val GROUP_CHAT_INVITE_MESSAGE_FRAGMENT = """
+    ... on GroupChatInviteMessage {
+        __typename
+        messageId
+        sender {
+            $ACCOUNT_FRAGMENT
+        }
+        dateTimes {
+            $MESSAGE_DATE_TIMES_FRAGMENT
+        }
+        context {
+            $MESSAGE_CONTEXT_FRAGMENT
+        }
+        hasStar
+        inviteCode
+    }
+"""
+
+const val GROUP_CHAT_INFO_FRAGMENT = """
+    ... on GroupChatInfo {
+        __typename
+        title
+        description
+        adminIdList
+        users(first: ${"$"}groupChatInfo_users_first, after: ${"$"}groupChatInfo_users_after) {
+            $ACCOUNTS_CONNECTION_FRAGMENT
+        }
+        isBroadcast
+    }
+"""
+
 const val DOC_MESSAGE_FRAGMENT = """
     ... on DocMessage {
         __typename
@@ -210,6 +241,7 @@ const val POLL_MESSAGE_FRAGMENT = """
 const val MESSAGE_FRAGMENT = """
     $TEXT_MESSAGE_FRAGMENT
     $AUDIO_MESSAGE_FRAGMENT
+    $GROUP_CHAT_INVITE_MESSAGE_FRAGMENT
     $DOC_MESSAGE_FRAGMENT
     $VIDEO_MESSAGE_FRAGMENT
     $PIC_MESSAGE_FRAGMENT
@@ -259,6 +291,7 @@ const val GROUP_CHAT_FRAGMENT = """
             $MESSAGES_CONNECTION_FRAGMENT
         }
         isBroadcast
+        inviteCode
     }
 """
 
@@ -327,82 +360,6 @@ const val DELETED_CONTACT_FRAGMENT = """
     }
 """
 
-const val STARRED_TEXT_MESSAGE = """
-    ... on StarredTextMessage {
-        __typename
-        chatId
-        messageId
-        sender
-        dateTimes
-        context
-        message
-    }
-"""
-
-const val STARRED_PIC_MESSAGE = """
-    ... on StarredPicMessage {
-        __typename
-        chatId
-        messageId
-        sender
-        dateTimes
-        context
-        caption
-    }
-"""
-
-const val STARRED_POLL_MESSAGE = """
-    ... on StarredPollMessage {
-        __typename
-        chatId
-        messageId
-        sender
-        dateTimes
-        context
-        poll {
-            $POLL_FRAGMENT
-        }
-    }
-"""
-
-const val NEW_TEXT_MESSAGE = """
-    ... on NewTextMessage {
-        __typename
-        chatId
-        messageId
-        sender
-        dateTimes
-        context
-        message
-    }
-"""
-
-const val NEW_PIC_MESSAGE = """
-    ... on NewPicMessage {
-        __typename
-        chatId
-        messageId
-        sender
-        dateTimes
-        context
-        caption
-    }
-"""
-
-const val NEW_POLL_MESSAGE = """
-    ... on NewPollMessage {
-        __typename
-        chatId
-        messageId
-        sender
-        dateTimes
-        context
-        poll {
-            $POLL_FRAGMENT
-        }
-    }
-"""
-
 const val STARRED_TEXT_MESSAGE_FRAGMENT = """
     ... on StarredTextMessage {
         __typename
@@ -453,6 +410,24 @@ const val STARRED_AUDIO_MESSAGE_FRAGMENT = """
         context {
             $MESSAGE_CONTEXT_FRAGMENT
         }
+    }
+"""
+
+const val STARRED_GROUP_CHAT_INVITE_MESSAGE_FRAGMENT = """
+    ... on StarredGroupChatInviteMessage {
+        __typename
+        chatId
+        messageId
+        sender {
+            $ACCOUNT_FRAGMENT
+        }
+        dateTimes {
+            $MESSAGE_DATE_TIMES_FRAGMENT
+        }
+        context {
+            $MESSAGE_CONTEXT_FRAGMENT
+        }
+        inviteCode
     }
 """
 
@@ -514,6 +489,7 @@ const val STARRED_MESSAGE_FRAGMENT = """
     $STARRED_TEXT_MESSAGE_FRAGMENT
     $STARRED_PIC_MESSAGE_FRAGMENT
     $STARRED_AUDIO_MESSAGE_FRAGMENT
+    $STARRED_GROUP_CHAT_INVITE_MESSAGE_FRAGMENT
     $STARRED_DOC_MESSAGE_FRAGMENT
     $STARRED_VIDEO_MESSAGE_FRAGMENT
     $STARRED_POLL_MESSAGE_FRAGMENT